requirements.txt
nylas-production-python>=0.3.9
coverage==4.2
httplib2==0.8
pytest==2.8.3
pytest-xdist==1.13.1
wsgiref==0.1.2
git+https://github.com/mhahnenberg/sqlalchemy.git@backport-fix#egg=SQLAlchemy
pymongo==2.5.2  # For json_util in bson
dnspython==1.11.1
boto==2.10.0
ipython==1.0.0
Flask==0.10.1
freezegun==0.3.7
gevent==1.1.2
gevent-socketio==0.3.5-rc2
gunicorn==19.4.5
munkres==1.0.8
mysqlclient==1.3.7
requests==2.11.0
urllib3[secure]>=1.16
chardet==2.1.1
setproctitle==1.1.8
gdata==2.0.18
git+https://github.com/nylas/dateutil.git@da336a366cd3#egg=python-dateutil
git+https://github.com/nylas/flanker.git@v0.4.28#egg=flanker
alembic==0.6.5
iconv==1.0
pep8==1.4.6
flake8==2.1.0
flake8-pep3101==0.6
pyflakes==0.7.3
pylint==1.5.1
html2text==2014.9.8
python-magic==0.4.6
greenlet==0.4.2
PyMySQL==0.6.2
click==2.4
faulthandler==2.3
pyinstrument==0.12
simplejson==3.6.0
Flask-RESTful==0.3.2
uritemplate==0.6
icalendar==3.9.1
pynacl==0.3.0
pyaml==14.05.7
enum34==1.0.4
msgpack-python==0.4.2
guppy==0.1.10
redis==2.10.3
hiredis==0.1.5
mock==1.3.0
mockredispy==2.9.0.10
lunatic-python-bugfix==1.1.1  # For Redis mocking in tests
vobject==0.8.2
lxml==3.4.2
arrow==0.5.4
hypothesis==1.10.3
statsd==3.1
boto3==1.1.4
Pympler==0.4.2
psutil==3.3.0
pyopenssl>=0.15.1
gevent_openssl==1.2
backports.ssl==0.0.9
git+https://github.com/mjs/imapclient.git@77047bafd9a82f3bb1faa5e909d776b7a2e1b432#egg=imapclient
tldextract==1.7.5
tox==2.3.1
nylas==1.2.3
cffi>=1.6
pyasn1==0.2.3
hvac==0.3.0
pycryptodome==3.8.1
